Fix TODO comments. Move custom upload function to a separate file.
