+++ stories/drag-n-drop-upload/src/DragNDropImageEditorUpload.tsx
@@ -7,6 +7,7 @@ import createFocusPlugin from '@draft-js-plugins/focus';
 import createResizeablePlugin from '@draft-js-plugins/resizeable';
 import createDragNDropUploadPlugin from '@draft-js-plugins/drag-n-drop-upload';
 import editorStyles from './editorStyles.css';
+import { customUpload } from './customUpload';
 
 // Plugins
 const blockDndPlugin = createBlockDndPlugin();
@@ -18,40 +19,6 @@ const decorator = composeDecorators(
   resizeablePlugin.decorator
 );
 const imagePlugin = createImagePlugin({ decorator });
-
-const FILE_SRC = '/images/canada-landscape-small.jpg';
-
-/*
-*
-* @name: handleUpload
-* @desc: Custom file upload function.
-* @param {param type} data - (){})
-* @param {param type} success - successful upload
-* @param {param type} failed - failed to upload files.
-* @param {param type} progress - handle progress.
-*/
-const customUpload = (data, success, failed, progress):void => {
-
- const mockResult = data.files.map(f => ({ name: f.name, src: FILE_SRC }));
-
- let intervalId = -1;
- let currentProgress = 0;
-
- intervalId = setInterval(() => {
-
-   if(currentProgress < 100) {
-      currentProgress+=10;
-      progress(currentProgress, mockResult[0].name);
-    }
-
-    if(currentProgress === 100) {
-      clearInterval(intervalId);
-      success(mockResult/*, { retainSrc: true }*/);
-    }
-
- }, 1000);
-
-}
 
 const dndFileUploadPlugin = createDragNDropUploadPlugin({
   handleUpload: customUpload,
